[김규섭] sprint7

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicMessageService.java
@@ -8,6 +8,10 @@ import com.sprint.mission.discodeit.entity.BinaryContent;
 import com.sprint.mission.discodeit.entity.Channel;
 import com.sprint.mission.discodeit.entity.Message;
 import com.sprint.mission.discodeit.entity.User;
+import com.sprint.mission.discodeit.exception.channel.ChannelNotFoundException;
+import com.sprint.mission.discodeit.exception.message.MessageNotFoundException;
+import com.sprint.mission.discodeit.exception.user.UserNotFoundException;
+import com.sprint.mission.discodeit.log.LogUtils;
 import com.sprint.mission.discodeit.mapper.MessageMapper;
 import com.sprint.mission.discodeit.mapper.PageResponseMapper;
 import com.sprint.mission.discodeit.repository.BinaryContentRepository;
@@ -16,12 +20,11 @@ import com.sprint.mission.discodeit.repository.MessageRepository;
 import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.service.MessageService;
 import com.sprint.mission.discodeit.storage.BinaryContentStorage;
-import jakarta.validation.Valid;
 import java.time.Instant;
 import java.util.List;
-import java.util.NoSuchElementException;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
@@ -31,6 +34,7 @@ import org.springframework.validation.annotation.Validated;
 @Service("messageService")
 @RequiredArgsConstructor
 @Validated
+@Slf4j
 public class BasicMessageService implements MessageService {
 
   private final MessageRepository messageRepository;
@@ -43,12 +47,15 @@ public class BasicMessageService implements MessageService {
 
   @Override
   @Transactional
-  public MessageDto create(@Valid MessageCreateCommand command) {
+  public MessageDto create(MessageCreateCommand command) {
+    log.debug("[BasicMessageService#create] try command={}", command.forLog());
+
     String content = command.content();
     User author = userRepository.findById(command.authorId())
-        .orElseThrow(() -> new NoSuchElementException("user not found :" + command.authorId()));
+        .orElseThrow(() -> new UserNotFoundException().addDetail("author", command.authorId()));
     Channel channel = channelRepository.findById(command.channelId())
-        .orElseThrow(() -> new NoSuchElementException("channel not found :" + command.channelId()));
+        .orElseThrow(
+            () -> new ChannelNotFoundException().addDetail("channel", command.channelId()));
 
     List<BinaryContent> attachments = command.attachments().stream()
         .map(request -> {
@@ -59,23 +66,27 @@ public class BasicMessageService implements MessageService {
           );
           binaryContentRepository.save(binaryContent);
           binaryContentStorage.put(binaryContent.getId(), request.bytes());
+          log.debug(
+              "[BasicMessageService#create] BinaryContent created: filename={}, contentType={}, size={}",
+              binaryContent.getFileName(),
+              binaryContent.getContentType(),
+              LogUtils.humanReadableSize(binaryContent.getSize()));
           return binaryContent;
         })
         .toList();
 
     Message message = new Message(content, channel, author, attachments);
 
-    return messageMapper.toDto(messageRepository.save(message));
+    MessageDto dto = messageMapper.toDto(messageRepository.save(message));
+    log.info("[MessageService#create] Message Created:{}", dto);
+
+    return dto;
   }
 
   @Override
   @Transactional(readOnly = true)
   public MessageDto findById(UUID messageId) {
-    return messageMapper.toDto(
-        messageRepository.findById(messageId)
-            .orElseThrow(
-                () -> new NoSuchElementException(
-                    "findById : 메세지를 찾을 수 없습니다. [" + messageId + "]")));
+    return messageMapper.toDto(validateId(messageId));
   }
 
   @Override
@@ -84,7 +95,7 @@ public class BasicMessageService implements MessageService {
       Pageable pageable) {
 
     if (!channelRepository.existsById(channelId)) {
-      throw new NoSuchElementException("findAllByChannelId : 채널을 찾을 수 없습니다. [" + channelId + "]");
+      throw new ChannelNotFoundException().addDetail("channel", channelId);
     }
 
     Slice<MessageDto> slice = (cursor == null)
@@ -103,25 +114,36 @@ public class BasicMessageService implements MessageService {
 
   @Override
   @Transactional
-  public MessageDto update(UUID messageId, @Valid MessageUpdateRequest messageUpdateRequest) {
-    Message message = messageRepository.findById(messageId)
-        .orElseThrow(
-            () -> new NoSuchElementException("update : 메세지를 찾을 수 없습니다. [" + messageId + "]"));
-    message.update(messageUpdateRequest.newContent());
+  public MessageDto update(UUID messageId, MessageUpdateRequest request) {
+    log.debug("[BasicMessageService#update] try messageId={} request={}", messageId, request);
+    Message message = validateId(messageId);
+    message.update(request.newContent());
 
-    return messageMapper.toDto(messageRepository.save(message));
+    MessageDto dto = messageMapper.toDto(messageRepository.save(message));
+    log.info("[MessageService#update] Message Updated:{}", dto);
+
+    return dto;
   }
 
   @Override
   @Transactional
   public void delete(UUID messageId) {
-    Message message = messageRepository.findById(messageId)
-        .orElseThrow(
-            () -> new NoSuchElementException("delete : 메세지를 찾을 수 없습니다. [" + messageId + "]"));
+    log.debug("[BasicMessageService#delete] try messageId={}", messageId);
+    Message message = validateId(messageId);
 
-    for (BinaryContent binaryContent : message.getAttachments()) {
-      binaryContentRepository.deleteById(binaryContent.getId());
+    if (message.getAttachments() != null && !message.getAttachments().isEmpty()) {
+      for (BinaryContent binaryContent : message.getAttachments()) {
+        binaryContentRepository.deleteById(binaryContent.getId());
+        log.debug("[MessageService#delete] BinaryContent Deleted:{}", binaryContent.getId());
+      }
     }
+
     messageRepository.deleteById(message.getId());
+    log.info("[MessageService#delete] Message Deleted:{}", message.getId());
+  }
+
+  private Message validateId(UUID messageId) {
+    return messageRepository.findById(messageId)
+        .orElseThrow(() -> new MessageNotFoundException().addDetail("id", messageId));
   }
 }
